3.2.4 - MaximumAdvertisingRevenue solution O(n²) + fix int overflow changed to long long

+++ algorithmic-toolbox/Assignment-3_GreedyAlgorithms/3.2.4-MaximumAdvertisingRevenue/MaxAdRevenue.h
@@ -3,7 +3,7 @@
 
 #include <vector>
 
-int maxAdRevenue(
+long long maxAdRevenue(
     const std::vector<int> & price,
     const std::vector<int> & clicks
 );
